# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api ONG",
    "version": "1.0.0",
    "description": "Exemplo de API para gerenciar o cadastro de ongs e gerar doações"
  },
  "paths": {
    "/ongs": {
      "get": {
        "summary": "Obter todas as ONGs",
        "tags": ["Rotas relacionadas Ongs"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OngList"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Criar nova ONG",
        "tags": ["Rotas relacionadas Ongs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OngRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ong criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OngResponse"
                }
              }
            }
          },
          "400": {
            "description": "Campo obrigatório não informado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "summary": "Atualizar ONG existente",
        "tags": ["Rotas relacionadas Ongs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OngRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ong atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OngResponse"
                }
              }
            }
          },
          "400": {
            "description": "Campo obrigatório não informado"
          },
          "404": {
            "description": "Ong não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Deleta uma ONG existente",
        "tags": ["Rotas relacionadas Ongs"],
        "requestBody": {
          "description": "ID da ONG a ser deletada",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ong deletada com sucesso, id: 1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "campo obrigatório não informado"
          },
          "404": {
            "description": "Ong não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/ong/doacoes": {
      "get": {
        "summary": "Obter doações de uma ONG ou todas as doações",
        "tags": ["Rotas relacionadas Doações"],
        "parameters": [
          {
            "in": "query",
            "name": "ongId",
            "schema": {
              "type": "integer"
            },
            "description": "ID da ONG para buscar doações específicas",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de doações ou total de doações da ONG",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Doacoes"
                    },
                    {
                      "$ref": "#/components/schemas/TotalDoacoes"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "ONG não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Inserir uma nova doação",
        "tags": ["Rotas relacionadas Doações"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "doador": {
                    "type": "string",
                    "description": "Nome do doador",
                    "example": "João Silva"
                  },
                  "valor": {
                    "type": "number",
                    "format": "float",
                    "description": "Valor da doação",
                    "example": 50.0
                  },
                  "ongId": {
                    "type": "integer",
                    "description": "Identificador único da ONG que receberá a doação",
                    "example": 1
                  }
                },
                "required": ["doador", "valor", "ongId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Doação criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "doação de R$ 50.00 para a Ong: ONG Amor Animal"
                    }
                  },
                  "required": ["mensagem"]
                }
              }
            }
          },
          "400": {
            "description": "Campos obrigatórios ausentes"
          },
          "404": {
            "description": "ONG não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OngRequest": {
        "type": "object",
        "properties": {
          "cnpj": {
            "type": "string",
            "description": "CNPJ da ONG",
            "example": "32979648000148"
          },
          "nome": {
            "type": "string",
            "description": "Nome da ONG",
            "example": "ONG Amor Animal"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo da ONG",
            "example": "Proteção Animal"
          },
          "data_criacao": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação da ONG",
            "example": "2023/04/20"
          },
          "desc": {
            "type": "string",
            "description": "Descrição da ONG",
            "example": "ONG que busca cuidar de animais abandonados"
          }
        },
        "required": ["cnpj", "nome", "tipo", "data_criacao", "desc"]
      },
      "OngResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID da ONG",
            "example": 1
          },
          "cnpj": {
            "type": "string",
            "description": "CNPJ da ONG",
            "example": "32979648000148"
          },
          "nome": {
            "type": "string",
            "description": "Nome da ONG",
            "example": "ONG Amor Animal"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo da ONG",
            "example": "Proteção Animal"
          },
          "data_criacao": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação da ONG",
            "example": "2023-04-20T03:00:00.000Z"
          },
          "desc": {
            "type": "string",
            "description": "Descrição da ONG",
            "example": "ONG que busca cuidar de animais abandonados"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de inclusão do registro da ONG no banco de dados",
            "example": "2023-04-20T19:42:24.455Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data do último update do registro da ONG no banco de dados",
            "example": "2023-04-20T19:42:24.455Z"
          }
        },
        "required": ["cnpj", "nome", "tipo", "data_criacao", "desc"]
      },
      "OngList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "cnpj": {
              "type": "string",
              "example": "32979648000100"
            },
            "nome": {
              "type": "string",
              "example": "Meu cachorrinho"
            },
            "tipo": {
              "type": "string",
              "example": "fundação"
            },
            "data_criacao": {
              "type": "string",
              "format": "date-time",
              "example": "2023-04-19T03:00:00.000Z"
            },
            "desc": {
              "type": "string",
              "example": "resgate de cachorrinhos abandonados"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time",
              "example": "2023-04-20T12:53:35.000Z"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time",
              "example": "2023-04-20T12:53:35.000Z"
            },
            "deletedAt": {
              "type": "string",
              "format": "date-time",
              "nullable": true,
              "example": null
            }
          }
        }
      },
      "Doacoes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único da doação",
            "example": 1
          },
          "doador": {
            "type": "string",
            "description": "Nome do doador",
            "example": "Cristiano Oliveira"
          },
          "valor": {
            "type": "number",
            "format": "float",
            "description": "Valor da doação",
            "example": 50.0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de inclusão do registro da ONG no banco de dados",
            "example": "2023-04-20T19:42:24.455Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data do último update do registro da ONG no banco de dados",
            "example": "2023-04-20T19:42:24.455Z"
          },
          "ongId": {
            "type": "integer",
            "description": "Identificador único da ONG que recebeu a doação",
            "example": 1
          }
        }
      },
      "TotalDoacoes": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string",
            "example": "A Ong: ONG Amor Animal recebeu R$ 100.00 em doações"
          }
        },
        "required": ["mensagem"]
      }
    }
  }
}